requirements.txt
torch>=2.1
transformers>=4.41
datasets>=2.20
tokenizers>=0.19
accelerate>=0.33
pyyaml>=6.0
evaluate>=0.4
huggingface_hub>=0.23
